worked on sidebar
