--- Resources/views/Widget/order.info.html.twig
@@ -1,52 +1,0 @@
-{#{% if data.id is defined and data is not null %}#}
-    {#<div id="sidebar-contact-list" data-id="{{ data.id }}" class="show-icon grid-row m-bottom-40">#}
-        {#<h2 class="m-bottom-0 action">{{ data.fullName }}#}
-            {#<span class="blue-icon"><span class="fa-pencil"></span></span>#}
-        {#</h2>#}
-        {#<span class="block action">#}
-            {#{% if data.title is defined and data.position is defined %}#}
-                {#{{ [data.title, data.position]|join(', ') }}#}
-            {#{% elseif data.title is defined %}#}
-                {#{{ data.title }}#}
-            {#{% elseif data.position is defined %}#}
-                {#{{ data.position }}#}
-            {#{% endif %}#}
-        {#</span>#}
-    {#</div>#}
-
-    {#{% if data.address is defined %}#}
-        {#{% include 'SuluContactBundle:Widgets:Partials/address.partial.html.twig' with {'address': data.address} only %}#}
-    {#{% endif %}#}
-
-    {#{% if data.phone or data.fax  is not empty %}#}
-        {#{% include 'SuluContactBundle:Widgets:Partials/phoneFax.partial.html.twig' with {'phone': data.phone, 'fax': data.fax} only %}#}
-    {#{% endif %}#}
-
-    {#{% if data.email or data.url is not empty %}#}
-        {#{% include 'SuluContactBundle:Widgets:Partials/emailUrl.partial.html.twig' with {'email': data.email, 'url': data.url} only %}#}
-    {#{% endif %}#}
-
-    {#{% if data.company is defined %}#}
-        {#<div class="grid-row m-bottom-30">#}
-            {#<h5 class="label"><%=translate('contacts.sidebar.mainCompany')%></h5>#}
-            {#<h4 class="m-top-0 m-bottom-0 action" data-id="{{ data.company.id }}" id="main-account">{{ data.company.name }}</h4>#}
-            {#{% if data.company.address is defined %}#}
-                {#<span class="grey block">#}
-                    {#{% if data.company.address.city is defined and data.company.address.country is defined %}#}
-                        {#{{ [data.company.address.city, data.company.address.country]|join(', ') }}#}
-                    {#{% elseif data.company.address.city is defined %}#}
-                        {#{{ data.company.address.city }}#}
-                    {#{% elseif data.company.address.country is defined %}#}
-                        {#{{ data.company.address.country }}#}
-                    {#{% endif %}#}
-                {#</span>#}
-            {#{% endif %}#}
-            {#<span class="block">#}
-             {#<a class="action" href="mailto:{{ data.company.email }}" target="_blank">#}
-                 {#{{ data.company.email }}#}
-             {#</a>#}
-         {#</span>#}
-        {#</div>#}
-    {#{% endif %}#}
-{#{% endif %}#}
-
